⚡️  : refactor response from helper to controller
